requirements.txt
discord.py
python-dotenv
asyncpg
aiohttp
beautifulsoup4
pydantic
pydantic-settings
Pillow